Crate wafel_layout

Defines a mapping from global variables to types/values.

A DataLayout is a mapping containing:

Type definitions

Global variable types

Constant values

This layout is used for accessing global data in Wafel. It can be constructed manually or read automatically from the DWARF debugging info of a DLL using DllLayout::read. The layouts for SM64 on the N64 are available using load_sm64_n64_layout.

Some SM64 specific data cannot be parsed from DWARF but is useful to have in the layout. To include these, use DataLayout::add_sm64_extras.

A json representation of this layout can be produced using the sm64_layout executable.

Structs

Constant

A constant’s value and source.

DataLayout

A description of accessible variables and types.

DllLayout

Debugging and structural information extracted from a DLL.

DllLayoutError

DllSegment

A segment defined in the DLL.

Global

A global variable or function.

Enums

ConstantSource

The source for a constant value.

DllLayoutErrorKind

LayoutLookupError

SM64LayoutError

Functions

append_dll_extension

Returns the DLL path, adding lib prefix or .so/.dll suffix if needed.

load_sm64_n64_layout

Return the layout for SM64 on the N64, including data added by DataLayout::add_sm64_extras.

read_dll_segments

Load segment definitions from the DLL.