Crate wafel_layout

Defines a mapping from global variables to types/values.

A DataLayout is a mapping containing:

Type definitions

Global variable types

Constant values

This layout is used for accessing global data in Wafel. It can be constructed manually or read automatically from the DWARF debugging info of a DLL using DllLayout::read.

Some SM64 specific data cannot be parsed from DWARF but is useful to have in the layout. To include these, use DataLayout::add_sm64_extras.

A json representation of this layout can be produced using the libsm64_layout executable.

Structs

Constant
A constant’s value and source.

DataLayout
A description of accessible variables and types.

DllLayout
Debugging and structural information extracted from a DLL.

DllLayoutError

DllSegment
A segment defined in the DLL.

Enums

ConstantSource
The source for a constant value.

DllLayoutErrorKind

LayoutLookupError

SM64ExtrasError

Functions

read_dll_segments
Load segment definitions from the DLL.